fix merge bugs.

+++ src/io/config.cpp
@@ -239,7 +239,7 @@ void OverallConfig::CheckParamConflict() {
   }
   // Check max_depth and num_leaves
   if (boosting_config.tree_config.max_depth > 0) {
-    int full_num_leaves = std::pow(2, boosting_config.tree_config.max_depth);
+    int full_num_leaves = static_cast<int>(std::pow(2, boosting_config.tree_config.max_depth));
     if (full_num_leaves > boosting_config.tree_config.num_leaves 
         && boosting_config.tree_config.num_leaves == kDefaultNumLeaves) {
       Log::Warning("Accuarcy may be bad since you didn't set num_leaves.");
